requirements.txt
flask==2.1.3
gunicorn==19.9.0
keras==2.9.0
tensorflow-cpu==2.9.0
numpy==1.23.3
albumentations==1.3.0
segmentation-models==1.0.1
pandas==1.5.2
